Refactor PipelinedSorter to be self-contained
